requirements.txt
streamlit
transformers
tortoise-tts

